# Utf8JsonReader: Do not allow non-standard line endings in single-line comments