--- a/src/System.Text.Json/src/Resources/Strings.resx
+++ b/src/System.Text.Json/src/Resources/Strings.resx
@@ -378,7 +378,7 @@
     <value>There is not enough data to read through the entire JSON array or object.</value>
   </data>
   <data name="UnexpectedEndOfLineSeparator" xml:space="preserve">
-    <value>Unexpected end of line separator while reading a comment.</value>
+    <value>Found invalid line or paragraph separator character while reading a comment.</value>
   </data>
   <data name="JsonSerializerDoesNotSupportComments" xml:space="preserve">
     <value>Comments cannot be stored when deserializing objects, only the Skip and Disallow comment handling modes are supported.</value>