--- a/src/System.Text.Json/src/System/Text/Json/ThrowHelper.cs
+++ b/src/System.Text.Json/src/System/Text/Json/ThrowHelper.cs
@@ -630,7 +630,7 @@ namespace System.Text.Json
         TrailingCommaNotAllowedBeforeObjectEnd,
         InvalidCharacterAtStartOfComment,
         UnexpectedEndOfDataWhileReadingComment,
-		UnexpectedEndOfLineSeparator,
+        UnexpectedEndOfLineSeparator,
         ExpectedOneCompleteToken,
         NotEnoughData,
     }