Fixes: https://github.com/dotnet/corefx/issues/33321
Fixes: https://github.com/dotnet/corefx/issues/37931
(also some other fixes - see below)

Some Unicode characters (i.e. U+2028, U+2029) are considered line endings and are sometimes used in some attack vectors when coming over the wire.

This change causes that parser will throw an exception if such character is encountered in a comment.

Also fixes couple of issues w.r.t to single-line comment handling with both mult- and single segment case around EOF/final block which I haven't reported. I.e.:
- end of json on `\r`
- end of json on empty comment in the end (i.e. `{}//` would cause parser to throw)

Perf results are here: dotnet/performance#543