@krwq 2019-05-31
@ahsonkhan I've rewrote SkipComment logic for multi-segment almost completely and have unified Consume/Skip. Haven't did anything similar for single-segment yet.