@krwq 2019-05-31
yes, this can wait, since there is currently no perf tests related to comments I'll let you know when I write them 😄