@krwq 2019-06-06
Perf results are here: https://github.com/dotnet/performance/pull/543
Multi-segment multiline comment specifically seems to be 2x faster for the common scenario, everything else is mostly the same. For some reason single-segment multiline comment regressed a bit (I have not changed that code path or anything related and have tried re-running tests)