@ahsonkhan 2019-06-08
> Multi-segment multiline comment specifically seems to be 2x faster for the common scenario, everything else is mostly the same. For some reason single-segment multiline comment regressed a bit (I have not changed that code path or anything related and have tried re-running tests)

I ran the perf tests locally and was not able to reproduce your results:

summary:
better: 1, geomean: 1.032
worse: 9, geomean: 1.073
total diff: 10

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.15 |           169.56 |           194.15 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.14 |           920.21 |          1048.72 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.08 |           118.35 |           127.87 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.07 |           143.18 |           152.90 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.07 |           172.37 |           183.75 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.05 |           150.26 |           157.17 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.04 |           142.63 |           148.78 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.04 |           147.42 |           153.37 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.04 |           121.93 |           126.25 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.03 |           131.92 |           127.82 |         |