@adamsitnik 2019-06-17
>  it looks like we are hitting a character limit on the ResultComparer output

I am not sure what would be the best way to solve it. The default output is console, which has a limited width. Most probably the easiest thing to do right now would be to shorten the ID.

```log
System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing
System.Text.Json.Tests.CommentsTests.Parsing
```

> can the ResultComparer store the result to disk somewhere

yes, you can pass the path to output csv file via `--csv $filePath` ([full docs](https://github.com/dotnet/performance/tree/master/src/tools/ResultsComparer))