@ahsonkhan 2019-06-18
> yes, you can pass the path to output csv file via `--csv $filePath`

> (by default - maybe md and csv) just like the actual perf runs (rather than us having to specify a path - maybe in one of the base/diff paths)

I am aware of the `--csv` option. I was asking, by default, without having to pass that option in.