@krwq 2019-06-21
@ahsonkhan ping, is there anything blocking here? I'm not personally concerned about tiny (< 20%) regression in perf here considering comments are not expected to be used in high perf scenarios and there have been couple of reliability fixes here