@ahsonkhan 2019-06-22
> is there anything blocking here? 

Nope, nothing blocking.

> I'm not personally concerned about tiny (< 20%) regression in perf here considering comments are not expected to be used in high perf scenarios and there have been couple of reliability fixes here

Agreed. I am still a bit concerned about the regression (especially for the "skip" mode), but we can work on perf optimizations separately.

Do you have the latest numbers based on your changes? We should double check if we have 10-20% regression on the "skip" case, because that is a bit too high.