@krwq 2019-06-26
@ahsonkhan seems the conflict is that we repeated this fix:
https://github.com/dotnet/corefx/pull/38036/files#diff-b12b49879ec5ef146f3373b69e66216bL3007

I'm building locally to double check my tests still build and run correctly after rebasing.